Final pass over revision tests
- remove the queueing_test go, which contained 1 test.
- remove duplicate constroller creation functions
- rename things
- make things check more things
- remove things that didn't make sense
